toolkit/components/satchel/test/test_popup_enter_event.html

<!DOCTYPE html>
<html>
<head>
  <meta charset="utf-8">
  <title>Test for events while the form history popup is open</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script src="/tests/SimpleTest/EventUtils.js"></script>
  <script type="text/javascript" src="satchel_common.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
Form History test: Test for events while the form history popup is open
<p id="display"></p>

<div id="content">
  <form id="form1">
    <input  type="text" name="field1">
    <button type="submit">Submit</button>
  </form>
</div>

<pre id="test">
<script class="testbody">
var form = document.getElementById("form1");
var input = getFormElementByName(1, "field1");
var expectedValue = "value1";

add_setup(async () => {
  await updateFormHistory([
    { op: "remove" },
    { op: "add", fieldname: "field1", value: "value1" },
  ]);
});

function handleEnter(evt) {
  if (evt.keyCode != KeyEvent.DOM_VK_RETURN) {
    return;
  }

  info("RETURN received for phase: " + evt.eventPhase);
  if (input.value == expectedValue) {
    ok(true, "RETURN should be received when the popup is closed");
    is(input.value, expectedValue, "Check input value when enter is pressed the 2nd time");
    info("form should submit with the default handler");
  } else {
    ok(false, "RETURN keypress shouldn't have been received when a popup item is selected");
  }
}

add_task(async function popupEnterEvent() {
  const submitTested = new Promise(resolve => {
    SpecialPowers.addSystemEventListener(input, "keypress", handleEnter, true);
    form.addEventListener("submit", evt => {
      evt.preventDefault();

      is(input.value, expectedValue, "Check input value in the submit handler");

      SpecialPowers.removeSystemEventListener(input, "keypress", handleEnter, true);

      resolve();
    }, { once: true });
  });
  const autocompleteShown = promiseACShown();

  // Focus the input before adjusting.value so that the caret goes to the end
  // (since OS X doesn't show the dropdown otherwise).
  input.focus();
  input.value = "value";
  input.focus();
  synthesizeKey("KEY_ArrowDown");
  await autocompleteShown;
  synthesizeKey("KEY_ArrowDown");
  synthesizeKey("KEY_Enter"); // select the first entry in the popup
  synthesizeKey("KEY_Enter"); // try to submit the form with the filled value
  await submitTested;
});

</script>
</pre>
</body>
</html>
